+++ CHANGELOG.md
@@ -21,7 +21,7 @@
 * adds option `default_op_type` to `opensearch_output` connector to set the default operation for indexing documents (default: index)
 * adds option `max_chunk_bytes` to `opensearch_output` connector to set the maximum size of the request in bytes (default: 100MB)
 * adds option `error_backlog_size` to logprep configuration to configure the queue size of the error queue
-
+* the opensearch default index is now only used for processed events, errors will be written to the error output, if configured
 
 ### Improvements